chore: add util to run multitx tests
